module1/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
bcrypt==4.0.1
email-validator==2.0.0
greenlet==3.0.1
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.11.1
google-generativeai==0.8.5
python-dotenv==0.21.0
